에러페이지 만들기, 에러컨트롤러, 커스텀익셉션, 인터셉터, 권한과 예외처리

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
+import shop.mtcoding.blog._core.errors.exception.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -41,29 +43,61 @@ public class BoardController {
 
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable Integer id, HttpServletRequest request){
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findById(id);
+
+
+        if(board ==null){
+            throw new Exception404("게시글을 찾을 수 없습니다");
+        }
         request.setAttribute("board", board);
         return "board/update-form";
     }
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id, BoardRequest.UpdateDTO reqDTO){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        //권한체크
+        if(sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 수정할 권한이 없습니다");
+        }
         boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
         return "redirect:/board/{id}";
     }
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        //권한체크
+        if(sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 삭제할 권한이 없습니다");
+        }
         boardRepository.deleteById(id);
         return "redirect:/";
     }
 
 
-
-
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board =boardRepository.findByIdJoinUser(id);
+
+        boolean isOwner=false;
+        if (sessionUser !=null){
+            if(sessionUser.getId()==board.getUser().getId()){
+                isOwner=true;
+            }
+        }
+
+        if(board ==null){
+            throw new Exception404("게시글을 찾을 수 없습니다");
+        }
+
+        request.setAttribute("isOwner", isOwner);
         request.setAttribute("board", board);
         return "board/detail";
     }
